requirements.txt
Pillow
numpy
scipy
matplotlib
opencv-contrib-python-headless